dev-requirements.in.txt
tensorflow_datasets
dvc[gcs]
dvc[gs]
isort
black==23.1.0
pyflakes
streamlit~=1.13.0
numpy==1.23.1
tensorflow~=2.11.0
tensorflow-serving-api~=2.11.0
tf-models-official
keras-cv~=0.4.0
pandas>=1.0.0
albumentations~=1.0.3
tqdm
matplotlib~=3.7.0
scikit-image~=0.19.3
Pillow~=9.4.0
scikit-learn~=1.2.1
pycocotools==2.0.4
click==8.1.3
requests==2.27.1
segmentation_models==1.0.1
distinctipy==1.1.5
google-cloud-storage>=2.5.0
google-api-python-client>=2.58.0
more-itertools~=8.14.0
opencv-python
dvclive==1.3.2
more-itertools~=8.14.0
ipython-autotime
